\name{SpeciesMixtures
}
\alias{SpeciesMixtures
}
\title{Package Description
}
\description{SpeciesMix is a package to fix finite mixture models to multispecies distributions. It implements the methods described in Dunstan et al. (2011) Ecological Modelling 222:955-963, doi:10.1016/j.ecolmodel.2010.11.030
}
\keyword{misc}
